lib: change to tshark for packet capture when run in an unshare
Tcpdump under fakeroot leaves residual processes when killed, i.e.,
killing the fakeroot process leaves both the child tcpdump and another
faked-sysv that we cannot control.

This workaround employs tshark for capture and tcpdump (in a fakeroot)
for reading .pcap files.  This because tshark turns out to be useless
for dumping payload (data) contents -- which is heavily relied upon in
our test cases.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>