suite/locked port: Add basic locked port tests
Locked port is used with 802.1x authentication, where a port should
not be accessible from a host behind a locked port unless the host
has been authenticated, whereby a static fdb entry is added to the
bridge port, allowing the host to traverse the port.

Tests include both with and without vlan filtering.