Support randomizing the topology
Specifying -r will cause brist to randomize the ordering and inversion
of the supplied pairs. In other words, expanding $b1 in a test
implementation does not necessarily result in the value that b1 was
assigned in the setup file. The process will of course still respect
the fact that the two ports making up a pair are fixed.

If an issue is discovered with a particular permutation, -R FILE will
cause brist to read in the permutation data for a previous test. The
typical usage is thus:

$ /usr/lib/brist/brist.sh -r
brist: suite ... log at /tmp/some-dir
...
FAIL: some-test
...
$ BRIST_TEST=some-test /usr/lib/brist/brist.sh -R /tmp/some-dir/shufdata

While not useful in a veth setting, on physical hardware, subtle
implementation bugs may only be exposed in particular configurations.

For example: Consider a 8-port switch made up of two cascaded switch
ASICs:

    CPU
     |
.----x----.   .---------.
|   sw0   y---z   sw1   |
'-1-2-3-4-'   '-5-6-7-8-'

You can imagine that a loop between 1 and 2 might act differently than
a loop between 4 and 5, due to the involvement of ports y and z.

Running repeated tests can also expose cleanup-related issues as the
same port will be used in different roles.