Add /tmp/brist-latest, a symlink pointing to the most recent run
Timestamping working dirs is good, because it means we don't override
any test output. On the flip side, this makes it hard to guess their
names. Autocompletion is also not great.

Compensate by maintaining a symlink that always points to the most
recent run. This makes sessions like this possible:

$ brist.sh -t some_bad_test
...
FAIL
$ tcpdump -r /tmp/brist-latest/some_bad_test/eth0.pcap
$ emacs some_bad_test.sh
$ brist.sh -t some_bad_test
...
FAIL
$ tcpdump -r /tmp/brist-latest/some_bad_test/eth0.pcap
$ emacs some_bad_test.sh
$ brist.sh -t some_bad_test
...
PASS