Fix sleep with fractions
Sleep with fractions (ie. `sleep 0.1`) is not supported by some
implementations of sleep(1) (notably Busybox's). This patch checks if
`sleep 0.1` is possible and if not monkey-patches sleep to use usleep(1).

Replacing sleep with usleep in all tests could be an alternative, but
usleep is not POSIX and may not be available. Also when not running in
verbose mode `sleep 0.1` will fail silently so new occurrences could
easily get committed without being noticed.

`set -e` could also be considered to catch these things earlier, but
that may have bigger impact.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>