mrouter_host: Add test
Generalize the existing mrouter_port test, such that the router port
can be either a bridge port, or the bridge itself.

This means we have to deal with all the nasty inconsistencies of
allowed flags for MDB entries etc. It also means that while we can use
the bridge's internal querier function for the port test, we must
generate our own queries when the querier is placed behind a bridge
port. For this reason, the host test requires that nemesis is
available on the test system.